Brought player hitstun in line with enemy hitstun
