+++ WhenPushComesToShove/Assets/3-Scripts/Player Scripts/PlayerHitstun.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 7468362213533ae4b9d3bfb93585f1e5
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
